requirements.txt
# build
protobuf

# test
pytest
pytest-cov
pylint
pydocstyle
flake8
mypy
GitPython

#docs
sphinx
sphinx_rtd_theme
